Created the Similar Recipes COmponenet, added it to search and category spotlight. Added Styles to Similar Recipes
