+++ src/Components/ReicpePage.tsx
@@ -23,10 +23,10 @@ interface ApiResponse {
   }
 
 //const KEY = "24ba6bf883a944a09e1f169a549f2c10";
-//const KEY = "cb61fb7dddc34daba2d7f61b391e90c1";
+const KEY = "cb61fb7dddc34daba2d7f61b391e90c1";
 //const KEY = "e3de5d36255e46babdbd21cbbbf5ec38";
 //const KEY = "fa0f376a27884351b9f852d1ae5e20f8";
-const KEY = "50980bc1ff884ed68509e87da2cf5db1";
+//const KEY = "50980bc1ff884ed68509e87da2cf5db1";
 
 const RecipePage = () => {
   const { mealType } = useParams<MatchParams>();
